Cleaner change role endpoint and delete role
Simplified the change_user_role function to use the list and not manually build the association, also decided to move it to users.py instead of roles.py as editing it from there makes more sense, and it fits with the see users by role endpoint. Also fixed delete role as suggested in review
